djangoapp/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==21.2.0
install==1.3.5
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymemcache==4.0.0
python-memcached==1.62
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
